I fixed an issue with where VisIt looks for private plugins when running a development version.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/branches/2.3RC@15066 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/common/state/MapNode.h
@@ -100,7 +100,7 @@ class STATE_API MapNode : public Variant
     void                 RemoveEntry(const std::string &);
     bool                 HasEntry(const std::string &) const;
     void                 GetEntryNames(stringVector &) const;
-    int                  GetNumEntries() const {return entries.size();}
+    int                  GetNumEntries() const {return (int)entries.size();}
     void                 Reset();
 
     virtual std::string  ToXML() const;
